Update README of Scalar DB and Scalar DL Auditor in Scalar DL Ledger branch
Since the README of Scalar DB and Scalar DL Auditor is updated by
`scripts/update-chart-docs.sh` in the Scalar DL Ledger branch, they are
always updated when we release new version of Scalar DL Ledger and we
need to exclude them every time.

To avoid the above unnecessary operation when we release Scalar DL Ledger
in the future, I updated the README of Scalar DB and Scalar DL Auditor
in the Scalar DL Ledger branch.

This commit doesn't affect release of Scalar DB and Scalar DL Auditor,
since I didn't update the chart version and app version in `Chart.yaml`.