Update README of Scalar DL Ledger in Scalar DL Auditor branch
Since the README of Scalar DL Ledger is updated by `scripts/update-chart-docs.sh`
in the Scalar DL Auditor branch, it is always updated when we release
new version of Scalar DL Auditor and we need to exclude it every time.

To avoid the above unnecessary operation when we release Scalar DL Auditor
in the future, I updated the README of Scalar DL Ledger in the Scalar
DL Auditor branch.

This commit doesn't affect release of Scalar DL Ledger, since I didn't
update the chart version and app version in `Chart.yaml`.